configure: unindex removed packages
For packages reported by xbps-checkvers as removed, generated makefile
finds files matching <pkgversion>.<arch>.xbps or
<pkgversion>.noarch.xbps. This is needed to know where is repodata
where package is indexed (path may contain multilib, aarch64, nonfree
subdirectories). Then, xbps-rindex is called to unindex package. Then,
package file is moved to backup/ subdirectory. If it is noarch, it is
recreated as empty file to allow it to be unindexed on other
architectures. Empty file, as not a valid package, won't be indexed again.